% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anidado.R
\name{anidado}
\alias{anidado}
\title{Diseño Anidado de factores fijos}
\usage{
anidado(y, ff, fa, data)
}
\arguments{
\item{y}{(vector) datos de la variable respuesta}

\item{ff}{(vector) datos del factor principal, que anida a otros}

\item{fa}{(vector) datos del factor anidado dentro del factor principal}

\item{data}{dataframe con los valores solicitados por los factores}
}
\value{
devuelve una tabla de analisis de varianza para el diseño anidado con factores fijos
}
\description{
Diseño Anidado de factores fijos
}
